[ResultPage] work_title/work_details_tableをwork_info_sectionに統合
